+++ cookbooks/arcgis-enterprise/attributes/default.rb
@@ -32,7 +32,7 @@ end
 default['arcgis']['run_as_msa'] = false
 default['arcgis']['run_as_user_auth_keys'] = nil
 
-default['arcgis']['version'] = '12.0'
+default['arcgis']['version'] = '12.1'
 
 default['arcgis']['cache_authorization_files'] = false
 default['arcgis']['configure_windows_firewall'] = false