+++ cookbooks/arcgis-repository/metadata.rb
@@ -4,8 +4,8 @@ maintainer_email 'contracts@esri.com'
 license          'Apache-2.0'
 description      'Downloads ArcGIS software setups from remote to local repositories'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '5.3.0'
-chef_version     '>= 15.3' if defined? chef_version
+version          '5.4.0'
+chef_version     '>= 15.3', '< 19'
 
 depends 'nfs', '~> 5.1'