+++ cookbooks/arcgis-video/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file is used to list changes made in each version of the arcgis-video cookbook.
 
+## 5.4.0
+
+- Added support for ArcGIS Video Server 12.1.
+- Added support for SSL certificates configuration for ArcGIS Video Server.
+
+## 5.3.0
+
+- Added support for ArcGIS Video Server 12.0.
+
 ## 5.2.0
 
 - Added support for ArcGIS Video Server 11.5.